academic_platform/pyproject.toml
[project]
name = "academic-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Yoma20",email = "98313300+Yoma20@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
